feat(mobile): enhance player UI with floating tab bar and control improvements
- Added PlayerTabBar component to display active track in the tab bar
- Replaced FloatingBar with PlayerTabBar for consistent player controls
- Improved play/pause button with animated transitions
- Updated progress bar text styling with tabular numbers
- Refined DismissIndicator styling in player screen
